Conflicts:
	website/traveldiary-project/public/javascripts/app.js
	website/traveldiary-project/public/javascripts/directives.js
	website/traveldiary-project/public/templates/vacation_details.html

+++ website/traveldiary-project/app/models/Keyword.java
@@ -5,14 +5,17 @@ package models;
  */
 
 import play.data.validation.Constraints;
-import javax.persistence.*;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 @Entity
 public class Keyword {
 
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
 
     @Constraints.Required
     private String keyword;
@@ -25,11 +28,11 @@ public class Keyword {
         this.keyword = keyword;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
