+++ website/traveldiary-project/app/models/Location.java
@@ -2,7 +2,9 @@ package models;
 
 import play.data.validation.Constraints;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 /**
  * Created by Rike on 06.06.2015.
@@ -12,7 +14,7 @@ public class Location {
 
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
 
     @Constraints.Required
     private String name;
@@ -25,11 +27,11 @@ public class Location {
     private String address;
     private String googleMapsLink;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
