+++ website/traveldiary-project/app/models/Role.java
@@ -15,15 +15,15 @@ public class Role {
 
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
     @Constraints.Required
     private String roleName;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -34,6 +34,5 @@ public class Role {
     public void setRoleName(String roleName) {
         this.roleName = roleName;
     }
-
-
+    
 }
